pyproject.toml
[tool.poetry]
name = "publish"
version = "1.0.0"
description = ""
authors = ["iwatos <nonbiriactive@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
mkdocs-material = "^5.1.1"
mkdocs = "^1.1"
mkdocs-roamlinks-plugin = "^0.1.3"
mkdocs-exclude = "^1.0.2"
GitPython = "^3.1.26"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"
pylint = "^2.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
